requirements.txt
anthropic>=0.40.0
pypdf>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
